requirements.txt
asgiref==3.3.1
Django==3.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-taggit==1.2.0
Markdown==3.2.1
Pillow==8.0.1
psycopg2-binary==2.8.4
pytz==2020.5
sqlparse==0.4.1
